Remove all known cyclic module dependencies
Breaks the five KNOWN_CYCLES from check-cyclic-dependencies.js:

- flycam_accessor <-> view_mode_accessor: inline the one-line
  getViewportRects state read (this cycle was part of the
  async_get_maximum_zoom_for_all_mags web worker's dependency graph)
- url_manager <-> model_initialization: extract applyState and friends
  into viewer/controller/apply_url_state.ts
- organization_cards <-> upgrade_plan_modal: extract the plan upgrade
  cards into admin/organization/plan_upgrade_cards.tsx
- dataset_table <-> folder_tree: extract the react-dnd item type into
  dashboard/advanced_dataset/dnd_types.ts
- plane -> plane_material_factory -> main_data_shaders.glsl -> plane:
  move PLANE_SUBDIVISION into viewer/constants.ts

The cycle checker now also uses the *.worker.ts files as entry points,
since they are only reachable via import.meta.glob / dynamic imports
which dpdm does not follow. This guards the web worker dependency
graphs against future cycles, where they are especially harmful.
KNOWN_CYCLES is now empty, so any new cycle fails the check.
