+++ frontend/javascripts/dashboard/folders/folder_tree.tsx
@@ -17,7 +17,7 @@ import { type Key, useCallback, useEffect, useRef, useState } from "react";
 import { type ConnectDropTarget, type DropTargetMonitor, useDrop } from "react-dnd";
 import type { FolderItem } from "types/api_types";
 import type { ArbitraryObject } from "types/type_utils";
-import { DraggableDatasetType } from "../advanced_dataset/dataset_table";
+import { DraggableDatasetType } from "../advanced_dataset/dnd_types";
 import {
   type DatasetCollectionContextValue,
   useDatasetCollectionContext,
